+++ tests/image_methods_test.py
@@ -839,7 +839,12 @@ def test_hald_clut(fx_asset):
         with Image(filename='hald:3') as hald:
             hald.gamma(0.367)
             img.hald_clut(hald)
-            assert was != img.signature
+        assert was != img.signature
+        was = img.signature
+        with Image(filename='hald:3') as hald:
+            hald.gamma(0.367)
+            img.hald_clut(hald, channel='red')
+        assert was != img.signature
 
 
 def test_implode(fx_asset):
@@ -1168,6 +1173,11 @@ def test_morphology_user_defined(fx_asset):
         img.morphology(method='dilate',
                        kernel='3x3: 0.3,0.6,0.3 0.6,1.0,0.6 0.3,0.6,0.3')
         assert was != img.signature
+        was = img.signature
+        img.morphology(method='dilate',
+                       kernel='3x3: 0.3,0.6,0.3 0.6,1.0,0.6 0.3,0.6,0.3',
+                       channel='red')
+        assert was != img.signature
         with raises(ValueError):
             img.morphology(method='dilate',
                            kernel='junk:0')
@@ -1277,6 +1287,11 @@ def test_opaque_paint():
         img.opaque_paint(target='white', fill='pink',
                          fuzz=0.25*img.quantum_range, invert=True)
         assert img[0, 0] == white
+    with Image(filename='WIZARD:') as img:
+        was = img.signature
+        img.opaque_paint(target='white', fill='pink',
+                         fuzz=0.25*img.quantum_range, channel='red')
+        assert was != img.signature
 
 
 def test_optimize_layers(fx_asset):
@@ -1601,6 +1616,9 @@ def test_selective_blur():
         was = img.signature
         img.selective_blur(8, 3, 0.1 * img.quantum_range)
         assert was != img.signature
+        was = img.signature
+        img.selective_blur(8, 3, 0.1 * img.quantum_range, channel='red')
+        assert was != img.signature
 
 
 def test_shade(fx_asset):
@@ -1634,9 +1652,6 @@ def test_sharpen(fx_asset):
         was = img.signature
         img.sharpen(radius=10.0, sigma=2.0)
         assert was != img.signature
-        was = img.signature
-        img.sharpen(radius=10.0, sigma=2.0, channel='red')
-        assert was != img.signature
         with raises(TypeError):
             img.sharpen(radius='hello')
         with raises(TypeError):
@@ -1667,12 +1682,6 @@ def test_sigmoidal_contrast():
         img.sigmoidal_contrast(sharpen=True,
                                strength=3,
                                midpoint=0.65 * img.quantum_range)
-        assert was != img.signature
-        was = img.signature
-        img.sigmoidal_contrast(sharpen=True,
-                               strength=3,
-                               midpoint=0.65 * img.quantum_range,
-                               channel='red')
         assert was != img.signature
 
 
@@ -1714,9 +1723,6 @@ def test_solarize():
         img.alpha_channel = 'off'  # Needed for IM-7
         img.solarize(0.5 * img.quantum_range)
         assert was != img.signature
-        was = img.signature
-        img.solarize(0.25 * img.quantum_range, channel='blue')
-        assert was != img.signature
 
 
 def test_sparse_color():
@@ -1754,9 +1760,6 @@ def test_statistic(fx_asset):
     with Image(filename='rose:') as img:
         was = img.signature
         img.statistic('median', 5, 5)
-        assert was != img.signature
-        was = img.signature
-        img.statistic('maximum', 5, 5, channel='red')
         assert was != img.signature
 
 
@@ -2015,10 +2018,6 @@ def test_unsharp_mask(fx_asset, display):
         assert 0.89 <= after.red <= 0.90
         assert 0.82 <= after.green <= 0.83
         assert 0.72 <= after.blue < 0.74
-    with Image(filename='rose:') as img:
-        was = img.signature
-        img.unsharp_mask(1.1, 1, 0.5, 0.001, channel='red')
-        assert was != img.signature
 
 
 def test_vignette(fx_asset):